Build using ECL
ECL does not build a single executable when attempting to build a
package-inferred-system using make-build. Instead it creates a directory
containing object archives and fasb files for each source file respectively. It
doesn't appear to overwrite these if the directory already exists, complicating
subsequent builds.

ASDF must be listed as a dependency, otherwise the resulting binary fails to
run.