Handle non-existent files when stat-ing
When building with ECL, the program fails while stat-ing file-paths containing
non-ascii characters (observed with umlauts). Such paths are not handled
correctly, resulting in queries about files that do not exist. This might be
caused by a bug in ECL's bundled asdf, as all paths are obtained via uiop and
the mangled names appear in the program output.